Completed call to Vehicle API
